requirements.txt
amqp==5.3.1
asgiref==3.9.1
beautifulsoup4==4.13.5
billiard==4.2.1
celery==5.5.3
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2.5
django-allauth==65.11.0
django-celery-beat==2.8.1
django-debug-toolbar==6.0.0
django-extensions==4.1
django-filter==25.1
django-tailwind==4.2.0
django-timezone-field==7.1
django_celery_results==2.6.0
djangorestframework==3.16.1
flower==2.0.1
humanize==4.13.0
kombu==5.5.4
Markdown==3.8.2
packaging==25.0
prometheus_client==0.22.1
prompt_toolkit==3.0.51
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.2.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
vine==5.1.0
wcwidth==0.2.13
